requirements.txt
flask
pandas
openpyxl
google-search-results
gunicorn
python-dotenv